modularize wallet send
